move ClipRect from tab box to tab bar
Tabs play an animation when they are opened and therefore need to be
clipped. Previously, the clipping was done directly by the tab box.
However, for performance reasons, and to eventually add a shadow to the
tabs for decorative purposes, the ClipRect has now been moved to the tab bar
instead.